Cleanup gNOI.
